+++ examples/scripts/compare_lithium_ion.py
@@ -26,7 +26,7 @@ models = [
 
 # load parameter values and process models and geometry
 param = models[0].default_parameter_values
-param["Current function [A]"] = 1.0
+param["Current function [A]"] = 1
 
 for model in models:
     param.process_model(model)
@@ -46,7 +46,7 @@ for model in models:
 
 # solve model
 solutions = [None] * len(models)
-t_eval = np.linspace(0, 0.3, 100)
+t_eval = np.linspace(0, 3600, 100)
 for i, model in enumerate(models):
     solutions[i] = model.default_solver.solve(model, t_eval)