+++ pybamm/simulation.py
@@ -5,6 +5,8 @@ import pickle
 import pybamm
 import numpy as np
 import copy
+import warnings
+import sys
 
 
 def isnotebook():
@@ -201,14 +203,12 @@ class Simulation:
                 )
 
             self._experiment_inputs.append(operating_inputs)
-            # Convert time to dimensionless
-            dt_dimensional = op[2]
-            if dt_dimensional is None:
+            # Add time to the experiment times
+            dt = op[2]
+            if dt is None:
                 # max simulation time: 1 week
-                dt_dimensional = 7 * 24 * 3600
-            tau = self._parameter_values.evaluate(self.model.timescale)
-            dt_dimensionless = dt_dimensional / tau
-            self._experiment_times.append(dt_dimensionless)
+                dt = 7 * 24 * 3600
+            self._experiment_times.append(dt)
 
         # add current and voltage events to the model
         # current events both negative and positive to catch specification
@@ -299,7 +299,7 @@ class Simulation:
         self._mesh = pybamm.Mesh(self._geometry, self._submesh_types, self._var_pts)
         self._disc = pybamm.Discretisation(self._mesh, self._spatial_methods)
         self._built_model = self._disc.process_model(
-            self._model, inplace=False, check_model=check_model
+            self._model_with_set_params, inplace=False, check_model=check_model
         )
 
     def solve(
@@ -317,10 +317,11 @@ class Simulation:
         Parameters
         ----------
         t_eval : numeric type, optional
-            The times at which to compute the solution. If None the model will
-            be solved for a full discharge (1 hour / C_rate) if the discharge
-            timescale is provided. Otherwise the model will be solved up to a
-            non-dimensional time of 1.
+            The times at which to compute the solution. If None and the parameter
+            "Current function [A]" is not read from data the model will
+            be solved for a full discharge (1 hour / C_rate). If None and the
+            parameter "Current function [A]" is read from data the model will be
+            solved at the times provided in the data.
         solver : :class:`pybamm.BaseSolver`
             The solver to use to solve the model.
         external_variables : dict
@@ -340,19 +341,67 @@ class Simulation:
             solver = self.solver
 
         if self.operating_mode == "without experiment":
-            # Solve the normal way, with a single solve
-            if t_eval is None:
+            # For drive cycles (current provided as data) we perform additional tests
+            # on t_eval (if provided) to ensure the returned solution captures the
+            # input. If the current is provided as data then the "Current function [A]"
+            # is the tuple (filename, data).
+            if isinstance(self._parameter_values["Current function [A]"], tuple):
+                filename = self._parameter_values["Current function [A]"][0]
+                time_data = self._parameter_values["Current function [A]"][1][:, 0]
+                # If no t_eval is provided, we use the times provided in the data.
+                if t_eval is None:
+                    pybamm.logger.info(
+                        "Setting t_eval as specified by the data '{}'".format(filename)
+                    )
+                    t_eval = time_data
+                # If t_eval is provided we first check if it contains all of the times
+                # in the data to within 10-12. If it doesn't, we then check
+                # that the largest gap in t_eval is smaller than the smallest gap in the
+                # time data (to ensure the resolution of t_eval is fine enough).
+                # We only raise a warning here as users may genuinely only want
+                # the solution returned at some specified points.
+                elif (
+                    set(np.round(time_data, 12)).issubset(set(np.round(t_eval, 12)))
+                ) is False:
+                    warnings.warn(
+                        """
+                        t_eval does not contain all of the time points in the data
+                        '{}'. Note: passing t_eval = None automatically sets t_eval
+                        to be the points in the data.
+                        """.format(
+                            filename
+                        ),
+                        pybamm.SolverWarning,
+                    )
+                    dt_data_min = np.min(np.diff(time_data))
+                    dt_eval_max = np.max(np.diff(t_eval))
+                    if dt_eval_max > dt_data_min + sys.float_info.epsilon:
+                        warnings.warn(
+                            """
+                            The largest timestep in t_eval ({}) is larger than
+                            the smallest timestep in the data ({}). The returned
+                            solution may not have the correct resolution to accurately
+                            capture the input. Try refining t_eval. Alternatively,
+                            passing t_eval = None automatically sets t_eval to be the
+                            points in the data.
+                            """.format(
+                                dt_eval_max, dt_data_min
+                            ),
+                            pybamm.SolverWarning,
+                        )
+            # If not using a drive cycle and t_eval is not provided, set t_eval
+            # to correspond to a single discharge
+            elif t_eval is None:
+                C_rate = self._parameter_values["C-rate"]
                 try:
-                    # Try to compute discharge time
-                    tau = self._parameter_values.evaluate(self.model.param.timescale)
-                    C_rate = self._parameter_values["C-rate"]
-                    t_end = 3600 / tau / C_rate
-                    t_eval = np.linspace(0, t_end, 100)
-                except AttributeError:
-                    t_eval = np.linspace(0, 1, 100)
+                    t_end = 3600 / C_rate
+                except TypeError:
+                    t_end = 3600
+                t_eval = np.linspace(0, t_end, 100)
 
             self.t_eval = t_eval
             self._solution = solver.solve(self.built_model, t_eval, inputs=inputs)
+
         elif self.operating_mode == "with experiment":
             if t_eval is not None:
                 pybamm.logger.warning(
@@ -367,11 +416,8 @@ class Simulation:
             ):
                 pybamm.logger.info(self.experiment.operating_conditions_strings[idx])
                 inputs.update(exp_inputs)
-                # Non-dimensionalise period
-                tau = self._parameter_values.evaluate(self.model.timescale)
-                freq = exp_inputs["period"] / tau
                 # Make sure we take at least 2 timesteps
-                npts = max(int(round(dt / freq)) + 1, 2)
+                npts = max(int(round(dt / exp_inputs["period"])) + 1, 2)
                 self.step(
                     dt, npts=npts, external_variables=external_variables, inputs=inputs
                 )
@@ -428,25 +474,15 @@ class Simulation:
         if solver is None:
             solver = self.solver
 
-        if save is False:
-            # Don't pass previous solution
-            self._solution = solver.step(
-                None,
-                self.built_model,
-                dt,
-                npts=npts,
-                external_variables=external_variables,
-                inputs=inputs,
-            )
-        else:
-            self._solution = solver.step(
-                self._solution,
-                self.built_model,
-                dt,
-                npts=npts,
-                external_variables=external_variables,
-                inputs=inputs,
-            )
+        self._solution = solver.step(
+            self._solution,
+            self.built_model,
+            dt,
+            npts=npts,
+            external_variables=external_variables,
+            inputs=inputs,
+            save=save,
+        )
 
     def get_variable_array(self, *variables):
         """