+++ pybamm/solvers/base_solver.py
@@ -48,7 +48,6 @@ class BaseSolver(object):
         self.max_steps = max_steps
 
         self.models_set_up = set()
-        self.model_step_times = {}
 
         # Defaults, can be overwritten by specific solver
         self.name = "Base solver"
@@ -115,7 +114,10 @@ class BaseSolver(object):
 
         """
         inputs = inputs or {}
-        y0 = model.concatenated_initial_conditions
+        y0 = model.concatenated_initial_conditions.evaluate(0, None, inputs)
+
+        # Set model timescale
+        model.timescale_eval = model.timescale.evaluate(u=inputs)
 
         # Check model.algebraic for ode solvers
         if self.ode_solver is True and len(model.algebraic) > 0:
@@ -135,9 +137,7 @@ class BaseSolver(object):
         if model.convert_to_format != "casadi":
             simp = pybamm.Simplification()
             # Create Jacobian from concatenated rhs and algebraic
-            y = pybamm.StateVector(
-                slice(0, np.size(model.concatenated_initial_conditions))
-            )
+            y = pybamm.StateVector(slice(0, np.size(y0)))
             # set up Jacobian object, for re-use of dict
             jacobian = pybamm.Jacobian()
         else:
@@ -216,17 +216,31 @@ class BaseSolver(object):
 
         # Check for heaviside functions in rhs and algebraic and add discontinuity
         # events if these exist.
-        # Note: only checks for the case of t < X, t <= X, X < t, or X <= t
-        for symbol in itertools.chain(model.concatenated_rhs.pre_order(),
-                                      model.concatenated_algebraic.pre_order()):
-            if isinstance(symbol, pybamm.Heaviside):
-                if symbol.right.id == pybamm.t.id:
-                    expr = symbol.left
-                elif symbol.left.id == pybamm.t.id:
-                    expr = symbol.right
+        # Note: only checks for the case of t < X, t <= X, X < t, or X <= t, but also
+        # accounts for the fact that t might be dimensional
+        # Only do this for DAE models as ODE models can deal with discontinuities fine
+        if len(model.algebraic) > 0:
+            for symbol in itertools.chain(
+                model.concatenated_rhs.pre_order(),
+                model.concatenated_algebraic.pre_order(),
+            ):
+                if isinstance(symbol, pybamm.Heaviside):
+                    # Dimensionless
+                    if symbol.right.id == pybamm.t.id:
+                        expr = symbol.left
+                    elif symbol.left.id == pybamm.t.id:
+                        expr = symbol.right
+                    # Dimensional
+                    elif symbol.right.id == (pybamm.t * model.timescale).id:
+                        expr = symbol.left.new_copy() / symbol.right.right.new_copy()
+                    elif symbol.left.id == (pybamm.t * model.timescale).id:
+                        expr = symbol.right.new_copy() / symbol.left.right.new_copy()
 
-                model.events.append(pybamm.Event(str(symbol), expr.new_copy(),
-                                                 pybamm.EventType.DISCONTINUITY))
+                    model.events.append(
+                        pybamm.Event(
+                            str(symbol), expr.new_copy(), pybamm.EventType.DISCONTINUITY
+                        )
+                    )
 
         # Process rhs, algebraic and event expressions
         rhs, rhs_eval, jac_rhs = process(model.concatenated_rhs, "RHS")
@@ -242,7 +256,8 @@ class BaseSolver(object):
         # discontinuity events are evaluated before the solver is called, so don't need
         # to process them
         discontinuity_events_eval = [
-            event for event in model.events
+            event
+            for event in model.events
             if event.event_type == pybamm.EventType.DISCONTINUITY
         ]
 
@@ -262,13 +277,13 @@ class BaseSolver(object):
             residuals, residuals_eval, jacobian_eval = process(all_states, "residuals")
             model.residuals_eval = residuals_eval
             model.jacobian_eval = jacobian_eval
-            y0_guess = model.concatenated_initial_conditions.flatten()
+            y0_guess = y0.flatten()
             model.y0 = self.calculate_consistent_state(model, 0, y0_guess)
         else:
             # can use DAE solver to solve ODE model
             model.residuals_eval = Residuals(rhs, "residuals", model)
             model.jacobian_eval = jac_rhs
-            model.y0 = model.concatenated_initial_conditions[:, 0]
+            model.y0 = y0.flatten()
 
         # Save CasADi functions for the CasADi solver
         # Note: when we pass to casadi the ode part of the problem must be in explicit
@@ -320,7 +335,7 @@ class BaseSolver(object):
             Initial conditions that are consistent with the algebraic equations (roots
             of the algebraic equations)
         """
-        pybamm.logger.info("Start calculating consistent initial conditions")
+        pybamm.logger.info("Start calculating consistent states")
         rhs = model.rhs_eval
         algebraic = model.algebraic_eval
         jac = model.jac_algebraic_eval
@@ -328,7 +343,7 @@ class BaseSolver(object):
             y0_guess = model.concatenated_initial_conditions.flatten()
 
         # Split y0_guess into differential and algebraic
-        len_rhs = rhs(0, y0_guess).shape[0]
+        len_rhs = rhs(time, y0_guess).shape[0]
         y0_diff, y0_alg_guess = np.split(y0_guess, [len_rhs])
 
         def root_fun(y0_alg):
@@ -336,8 +351,8 @@ class BaseSolver(object):
             y0 = np.concatenate([y0_diff, y0_alg])
             out = algebraic(time, y0)
             pybamm.logger.debug(
-                "Evaluating algebraic equations at t=0, L2-norm is {}".format(
-                    np.linalg.norm(out)
+                "Evaluating algebraic equations at t={}, L2-norm is {}".format(
+                    time * model.timescale, np.linalg.norm(out)
                 )
             )
             return out
@@ -425,6 +440,8 @@ class BaseSolver(object):
         if (np.diff(t_eval) < 0).any():
             raise pybamm.SolverError("t_eval must increase monotonically")
 
+        # Non-dimensionalise t_eval
+
         # Set up
         timer = pybamm.Timer()
 
@@ -433,6 +450,14 @@ class BaseSolver(object):
         inputs = inputs or {}
         ext_and_inputs = {**external_variables, **inputs}
 
+        # Raise warning if t_eval looks like it was supposed to be dimensionless
+        # already
+        if t_eval[-1] < 0.5:
+            raise pybamm.SolverError(
+                """It looks like t_eval might be dimensionless.
+                t_eval should now be provided in seconds"""
+            )
+
         # Set up (if not done already)
         if model not in self.models_set_up:
             self.set_up(model, ext_and_inputs)
@@ -440,6 +465,8 @@ class BaseSolver(object):
             self.models_set_up.add(model)
         else:
             set_up_time = 0
+        # Non-dimensionalise time
+        t_eval_dimensionless = t_eval / model.timescale_eval
         # Solve
         # Set inputs and external
         self.set_inputs(model, ext_and_inputs)
@@ -452,68 +479,77 @@ class BaseSolver(object):
 
         # make sure they are increasing in time
         discontinuities = sorted(discontinuities)
-        pybamm.logger.info(
-            'Discontinuity events found at t = {}'.format(discontinuities)
-        )
+
         # remove any identical discontinuities
         discontinuities = [
-            v for i, v in enumerate(discontinuities)
-            if i == len(discontinuities) - 1
-            or discontinuities[i] < discontinuities[i + 1]
+            v
+            for i, v in enumerate(discontinuities)
+            if (
+                i == len(discontinuities) - 1
+                or discontinuities[i] < discontinuities[i + 1]
+            )
+            and v > 0
         ]
+
+        if len(discontinuities) > 0:
+            pybamm.logger.info(
+                "Discontinuity events found at t = {}".format(discontinuities)
+            )
+        else:
+            pybamm.logger.info("No discontinuity events found")
 
         # insert time points around discontinuities in t_eval
         # keep track of sub sections to integrate by storing start and end indices
         start_indices = [0]
         end_indices = []
+        eps = sys.float_info.epsilon
         for dtime in discontinuities:
-            dindex = np.searchsorted(t_eval, dtime, side='left')
+            dindex = np.searchsorted(t_eval_dimensionless, dtime, side="left")
             end_indices.append(dindex + 1)
             start_indices.append(dindex + 1)
-            if t_eval[dindex] == dtime:
-                t_eval[dindex] += sys.float_info.epsilon
-                t_eval = np.insert(t_eval, dindex, dtime - sys.float_info.epsilon)
+            if dtime - eps < t_eval_dimensionless[dindex] < dtime + eps:
+                t_eval_dimensionless[dindex] += eps
+                t_eval_dimensionless = np.insert(
+                    t_eval_dimensionless, dindex, dtime - eps
+                )
             else:
-                t_eval = np.insert(t_eval, dindex,
-                                   [dtime - sys.float_info.epsilon,
-                                    dtime + sys.float_info.epsilon])
-        end_indices.append(len(t_eval))
+                t_eval_dimensionless = np.insert(
+                    t_eval_dimensionless, dindex, [dtime - eps, dtime + eps],
+                )
+        end_indices.append(len(t_eval_dimensionless))
 
-        # integrate separatly over each time segment and accumulate into the solution
+        # integrate separately over each time segment and accumulate into the solution
         # object, restarting the solver at each discontinuity (and recalculating a
         # consistent state afterwards if a dae)
         old_y0 = model.y0
         solution = None
         for start_index, end_index in zip(start_indices, end_indices):
-            pybamm.logger.info("Calling solver for {} < t < {}"
-                               .format(t_eval[start_index], t_eval[end_index - 1]))
+            pybamm.logger.info(
+                "Calling solver for {} < t < {}".format(
+                    t_eval_dimensionless[start_index] * model.timescale_eval,
+                    t_eval_dimensionless[end_index - 1] * model.timescale_eval,
+                )
+            )
             timer.reset()
+            new_solution = self._integrate(
+                model, t_eval_dimensionless[start_index:end_index], ext_and_inputs
+            )
+            new_solution.solve_time = timer.time()
             if solution is None:
-                solution = self._integrate(
-                    model, t_eval[start_index:end_index], ext_and_inputs)
-                solution.solve_time = timer.time()
+                solution = new_solution
             else:
-                new_solution = self._integrate(
-                    model, t_eval[start_index:end_index], ext_and_inputs)
-                new_solution.solve_time = timer.time()
                 solution.append(new_solution, start_index=0)
 
             if solution.termination != "final time":
                 break
 
-            if end_index != len(t_eval):
+            if end_index != len(t_eval_dimensionless):
                 # setup for next integration subsection
-                y0_guess = solution.y[:, -1]
-                if model.algebraic:
-                    model.y0 = self.calculate_consistent_state(
-                        model, t_eval[end_index], y0_guess)
-                else:
-                    model.y0 = y0_guess
-
                 last_state = solution.y[:, -1]
                 if len(model.algebraic) > 0:
                     model.y0 = self.calculate_consistent_state(
-                        model, t_eval[end_index], last_state)
+                        model, t_eval_dimensionless[end_index], last_state
+                    )
                 else:
                     model.y0 = last_state
 
@@ -525,7 +561,7 @@ class BaseSolver(object):
 
         # Add model and inputs to solution
         solution.model = model
-        solution.inputs = inputs
+        solution.inputs = ext_and_inputs
 
         # Identify the event that caused termination
         termination = self.get_termination_reason(solution, model.events)
@@ -541,7 +577,14 @@ class BaseSolver(object):
         return solution
 
     def step(
-        self, old_solution, model, dt, npts=2, external_variables=None, inputs=None
+        self,
+        old_solution,
+        model,
+        dt,
+        npts=2,
+        external_variables=None,
+        inputs=None,
+        save=True,
     ):
         """
         Step the solution of the model forward by a given time increment. The
@@ -565,7 +608,8 @@ class BaseSolver(object):
             values at the current time
         inputs : dict, optional
             Any input parameters to pass to the model when solving
-
+        save : bool
+            Turn on to store the solution of all previous timesteps
 
         Raises
         ------
@@ -595,19 +639,23 @@ class BaseSolver(object):
         ext_and_inputs = {**external_variables, **inputs}
 
         # Run set up on first step
-        if model not in self.model_step_times:
+        if old_solution is None:
             pybamm.logger.info(
                 "Start stepping {} with {}".format(model.name, self.name)
             )
             self.set_up(model, ext_and_inputs)
-            self.model_step_times[model] = 0.0
+            t = 0.0
             set_up_time = timer.time()
         else:
+            # initialize with old solution
+            t = old_solution.t[-1]
+            model.y0 = old_solution.y[:, -1]
             set_up_time = 0
 
+        # Non-dimensionalise dt
+        dt_dimensionless = dt / model.timescale_eval
         # Step
-        t = self.model_step_times[model]
-        t_eval = np.linspace(t, t + dt, npts)
+        t_eval = np.linspace(t, t + dt_dimensionless, npts)
         # Set inputs and external
         self.set_inputs(model, ext_and_inputs)
 
@@ -621,14 +669,10 @@ class BaseSolver(object):
 
         # Add model and inputs to solution
         solution.model = model
-        solution.inputs = inputs
+        solution.inputs = ext_and_inputs
 
         # Identify the event that caused termination
         termination = self.get_termination_reason(solution, model.events)
-
-        # Set self.t and self.y0 to their values at the final step
-        self.model_step_times[model] = solution.t[-1]
-        model.y0 = solution.y[:, -1]
 
         pybamm.logger.debug("Finish stepping {} ({})".format(model.name, termination))
         if set_up_time:
@@ -643,7 +687,7 @@ class BaseSolver(object):
             pybamm.logger.debug(
                 "Step time: {}".format(timer.format(solution.solve_time))
             )
-        if old_solution is None:
+        if save is False or old_solution is None:
             return solution
         else:
             return old_solution + solution
@@ -704,10 +748,16 @@ class SolverCallable:
             self.inputs = inputs
         elif self.form == "casadi":
             self.inputs = casadi.vertcat(*[x for x in inputs.values()])
+        self.timescale = self.model.timescale_eval
 
     def __call__(self, t, y):
         y = y[:, np.newaxis]
         if self.name in ["RHS", "algebraic", "residuals", "event"]:
+            pybamm.logger.debug(
+                "Evaluating {} for {} at t={}".format(
+                    self.name, self.model.name, t * self.timescale
+                )
+            )
             return self.function(t, y).flatten()
         else:
             return self.function(t, y)
@@ -731,8 +781,5 @@ class Residuals(SolverCallable):
         self.mass_matrix = model.mass_matrix.entries
 
     def __call__(self, t, y, ydot):
-        pybamm.logger.debug(
-            "Evaluating residuals for {} at t={}".format(self.model.name, t)
-        )
         states_eval = super().__call__(t, y)
         return states_eval - self.mass_matrix @ ydot