+++ tests/unit/test_solvers/test_base_solver.py
@@ -41,6 +41,7 @@ class TestBaseSolver(unittest.TestCase):
         model = pybamm.BaseModel()
         a = pybamm.Scalar(1)
         model.algebraic = {a: a}
+        model.concatenated_initial_conditions = pybamm.Scalar(0)
         solver = pybamm.ScipySolver()
         with self.assertRaisesRegex(pybamm.SolverError, "Cannot use ODE solver"):
             solver.set_up(model)
@@ -50,6 +51,7 @@ class TestBaseSolver(unittest.TestCase):
         class ScalarModel:
             concatenated_initial_conditions = np.array([[2]])
             jac_algebraic_eval = None
+            timescale = 1
 
             def rhs_eval(self, t, y):
                 return np.array([])
@@ -68,6 +70,7 @@ class TestBaseSolver(unittest.TestCase):
         class VectorModel:
             concatenated_initial_conditions = np.zeros_like(vec)
             jac_algebraic_eval = None
+            timescale = 1
 
             def rhs_eval(self, t, y):
                 return y[0:1]
@@ -101,6 +104,7 @@ class TestBaseSolver(unittest.TestCase):
         class Model:
             concatenated_initial_conditions = np.array([2])
             jac_algebraic_eval = None
+            timescale = 1
 
             def rhs_eval(self, t, y):
                 return np.array([])
@@ -122,6 +126,16 @@ class TestBaseSolver(unittest.TestCase):
             "Could not find consistent initial conditions: solver terminated",
         ):
             solver.calculate_consistent_state(Model())
+
+    def test_time_too_short(self):
+        solver = pybamm.BaseSolver()
+        model = pybamm.BaseModel()
+        v = pybamm.StateVector(slice(0, 1))
+        model.rhs = {v: v}
+        with self.assertRaisesRegex(
+            pybamm.SolverError, "It looks like t_eval might be dimensionless"
+        ):
+            solver.solve(model, np.linspace(0, 0.1))
 
 
 if __name__ == "__main__":